standby switchover: avoid version-specific repmgr function calls
A valid use-case for "repmgr standby switchover" is to promote an
upgraded standby to primary, and demote the current primary (running
an older repmgr version) so that in turn can be updated.

It should at this point be feasible to make the calls made from
the newer version (promotion candidate) compatible with the functionality
available on the the older version (demotion candidate), even if that
means some newer functionality (e.g. checking whether repmgrd is running
or not) cannot be used.

As a first step, avoid making function calls specific to a particular
repmgr version in the "get_replication_info()" function, as the
information returned by those is not required for a switchover anyway.

Per GitHub #579 and possibly GitHub #569.